requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
backoff==2.2.1
bidict==0.22.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
chainlit==0.7.604
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
Deprecated==1.2.14
dnspython==2.4.2
docx2txt==0.8
exceptiongroup==1.1.3
fastapi==0.100.1
fastapi-socketio==0.0.10
filetype==1.2.0
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
langchain==0.0.330
langsmith==0.0.57
Lazify==0.4.0
loguru==0.7.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.1
openai==0.28.1
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pinecone-client==2.2.4
protobuf==4.25.0
pyarrow==14.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-engineio==4.8.0
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
streamlit==1.28.1
syncer==2.0.3
tenacity==8.2.3
tiktoken==0.5.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uptrace==1.21.0
urllib3==2.0.7
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.20.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
